+++ CosmosToPostgreSQL/CosmosToPostgreSQL/CosmosTextResource.cs
@@ -13,6 +13,6 @@ namespace CosmosToPostgreSQL
     {
         [JsonProperty(PropertyName = "_ts")]
         [JsonPropertyName("_ts")]
-        public int Ts { get; set; }
+        public long Ts { get; set; }
     }
 }
